+++ __tests__/Testy-test.tsx
@@ -1,0 +1,10 @@
+/**
+ * @format
+ */
+
+import 'react-native';
+import { ScreenNames } from '../src/types';
+
+test('Default test', () => {
+  expect(ScreenNames.Login).toMatchInlineSnapshot(`"Login"`);
+});