+++ src/navigation/LoginStack.tsx
@@ -1,0 +1,4 @@
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
+/** The Stack Navigator for the Login Screen. */
+export const LoginStack = createNativeStackNavigator();