+++ src/types/index.ts
@@ -1,0 +1,8 @@
+/** 
+ * String names for the screens. 
+ * 
+ * Used in the stack navigators
+ */
+export enum ScreenNames {
+    Login = 'Login'
+}