+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 # v0.0.1
+- #2: Add React Navigation